py-sock/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
cachecontrol==0.12.5      # via firebase-admin
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask, pip-tools
dnspython==1.16.0         # via eventlet
eventlet==0.25.0
firebase-admin==2.17.0
flask-socketio==4.1.0
flask==1.1.1
google-api-core[grpc]==1.14.0  # via firebase-admin, google-cloud-core, google-cloud-firestore, google-cloud-logging
google-api-python-client==1.7.9  # via firebase-admin
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.3        # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-storage
google-cloud-core==1.0.2  # via google-cloud-firestore, google-cloud-logging, google-cloud-storage
google-cloud-firestore==1.3.0  # via firebase-admin
google-cloud-logging==1.11.0
google-cloud-storage==1.16.1  # via firebase-admin
google-resumable-media==0.3.2  # via google-cloud-storage
googleapis-common-protos==1.6.0  # via google-api-core
greenlet==0.4.15          # via eventlet
grpcio==1.22.0            # via google-api-core
gunicorn==19.9.0
httplib2==0.13.0          # via google-api-python-client, google-auth-httplib2
idna==2.8                 # via requests
importlib-metadata==0.18  # via pluggy, pytest
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask
markupsafe==1.1.1         # via jinja2
monotonic==1.5            # via eventlet
more-itertools==7.1.0     # via pytest
msgpack==0.6.1            # via cachecontrol
packaging==19.0           # via pytest
pip-tools==3.9.0
pluggy==0.12.0            # via pytest
protobuf==3.9.0           # via google-api-core, googleapis-common-protos
py==1.8.0                 # via pytest
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
pyparsing==2.4.0          # via packaging
pytest==5.0.1
python-dateutil==2.8.0
python-engineio==3.8.2.post1  # via python-socketio
python-socketio==4.2.0    # via flask-socketio
pytz==2019.1              # via google-api-core, google-cloud-firestore
redis==3.2.1
requests==2.22.0          # via cachecontrol, google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via eventlet, firebase-admin, google-api-core, google-api-python-client, google-auth, google-resumable-media, grpcio, packaging, pip-tools, protobuf, python-dateutil, python-engineio, python-socketio
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.25.3           # via requests
wcwidth==0.1.7            # via pytest
werkzeug==0.15.5          # via flask
zipp==0.5.2               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via google-api-core, protobuf
